added test case file containing test cases for given case study! also renamed some file names
