+++ naren/MySqlProject/app/controllers/answer_controller.js
@@ -1,5 +1,6 @@
-// FOR QUERYING THE DATABASE
-const sql = require("./db.js")
+// Establishing DataBase connection
+const db = require("../database/db.js")
+const dbConnection= db.getDbConnection()
 
 /*
 
@@ -7,9 +8,22 @@ const sql = require("./db.js")
 
 */
 
-// USE CASE 4 OF CASE STUDY -Part 1
-// POSTING A ANSWER
+// QUERY STRINGS FOR THIS CONTROLLER
+const authenticateUserQueryString = "SELECT * FROM UserDetails WHERE userName = ? AND password = ?;";
+const insertAnswerQueryString = "INSERT INTO Answers SET ?;";
+const getAnswerQueryString = "SELECT * FROM Answers WHERE userName = ? AND questionId = ?;";
+const updateAnswerQueryString = "UPDATE Answers SET content = ? WHERE userName =? AND questionId =?;";
+const getUserWhoPostedAnswerQueryString = "SELECT userName FROM Answers WHERE answerId = ?;";
+const insertUpvoteQueryString = "INSERT INTO Votes SET ?;";
+const insertUserPointsQueryString = "INSERT INTO UserPoints(userName) VALUES(?);"
+const deleteVotesQueryString = "DELETE FROM Votes WHERE answerId = ?;";
+const deleteAnswerQueryString = "DELETE FROM Answers WHERE answerId = ?;";
+const getAnswerDetailsQueryString = "SELECT * FROM Answers WHERE answerId = ? AND userName = ?;";
+const getUserPointsQueryString = "SELECT userName,Count(*) as Points FROM UserPoints GROUP BY userName ORDER BY Points DESC;";
+const getVoteCountQueryString = "SELECT Votes.answerId, Answers.content, Count(*) as VoteCount FROM Votes, Answers WHERE Votes.answerId = Answers.answerId GROUP BY Votes.answerId ORDER BY VoteCount DESC;";
+
 /*
+USE CASE 4 OF CASE STUDY -Part 1 : POSTING A ANSWER
 
 Function : postAnswer
 
@@ -41,53 +55,45 @@ Response of the API with HTTP STATUS codes:
 
 */
 
-const postAnswer = (req,res) =>
+const postAnswer = async (req,res) =>
 {
     const answersData =
     {
         userName : req.body.user_Details.userName,
         questionId : req.body.question.questionId,
         content : req.body.question.answer
-    }
+    };
     const userData =
     {
         userName : req.body.user_Details.userName,
         password : req.body.user_Details.password
-    }
-
-    // lets do the authentication part later
-    const authenticateQueryString = 'SELECT * FROM UserDetails where userName = ? and password =?;';
-    sql.query(authenticateQueryString,[userData.userName,userData.password],(err,result) => 
+    };
+    try
     {
-        if(err ||result.length == 0 )
+        var userAuthenticationDetails = await dbConnection.query(authenticateUserQueryString, [userData.userName, userData.password]);
+        if (userAuthenticationDetails.length == 0)
         {
-            console.log("Error: INVALID CREDENTIALS FOR LOGIN. PLEASE TRY AGAIN");
-            res.status(401).send({ message: "Error: INVALID CREDENTIALS FOR LOGIN. PLEASE TRY AGAIN"});
+            res.status(401).send({message: "Error: INVALID CREDENTIALS FOR LOGIN. PLEASE TRY AGAIN"});
+            return;
         }
-        else
+        try
         {
-            const queryString = 'INSERT INTO Answers SET ?;';
-            sql.query(queryString,answersData,(err,result) => 
-            {
-                if(err)
-                {
-                    console.log("error:maybe check for valid QUESTION ID OR IF user has already answered this question before. then try a PUT REQUEST TO update the answer",err);
-                    res.status(400).send({ message: 'This is an error! maybe check for valid QUESTION ID OR IF user has already answered this question before. then try a PUT REQUEST TO update the answer'});
-                }
-                else
-                {
-                    console.log("Answer for question "+answersData.questionId+" posted Successfully with ID "+result.insertId);
-                    res.status(201).send({message:"Answer for question "+answersData.questionId+" posted Successfully with ID "+result.insertId});
-                }
-            })
+        var insertAnswerQueryResult = await dbConnection.query(insertAnswerQueryString,answersData);
+        res.status(201).send({message: "Answer for posted Successfully with ID " + insertAnswerQueryResult.insertId});
         }
-    })
+        catch (err)
+        {
+            res.status(400).send({message: "Error: Check for valid QUESTION ID OR If user has already answered this question before. then try a PUT REQUEST TO update the answer"});
+        }
+    }
+    catch (err)
+    {
+        res.status(500).send({message: "This is an internal server error!"});
+    }        
 }   
 
-// USE CASE 4 OF CASE STUDY -Part 2
-// UPDATING AN ANSWER
 /*
-
+USE CASE 4 OF CASE STUDY -Part 2 : UPDATING AN ANSWER
 Function : updateAnswer
 
 API TYPE : PUT REQUEST
@@ -115,64 +121,45 @@ Response of the API with HTTP STATUS codes:
 500: "Internal Server Error "
 
 */
-const updateAnswer = (req,res) =>
+
+const updateAnswer = async (req,res) =>
 {
     const answersData =
     {
         userName : req.body.user_Details.userName,
         questionId : req.body.question.questionId,
         content : req.body.question.answer
-    }
+    };
     const userData =
     {
         userName : req.body.user_Details.userName,
         password : req.body.user_Details.password
-    }
-    const authenticateQueryString = 'SELECT * FROM UserDetails where userName = ? and password =?;';
-    sql.query(authenticateQueryString,[userData.userName,userData.password],(err,result) => 
+    };
+    try
     {
-        if(err ||result.length == 0 )
+        var userAuthenticationDetails = await dbConnection.query(authenticateUserQueryString, [userData.userName, userData.password]);
+        if (userAuthenticationDetails.length == 0)
         {
-            console.log("Error: INVALID CREDENTIALS FOR LOGIN. PLEASE TRY AGAIN");
-            res.status(401).send({ message: "Error: INVALID CREDENTIALS FOR LOGIN. PLEASE TRY AGAIN"});
+            res.status(401).send({message: "Error: INVALID CREDENTIALS FOR LOGIN. PLEASE TRY AGAIN"});
+            return;
         }
-        else
+        var questionUserDetails = await dbConnection.query(getAnswerQueryString, [answersData.userName, answersData.questionId]);
+        if (questionUserDetails == 0)
         {
-            const queryString = 'SELECT * FROM Answers where userName = ? and questionId = ?;';
-            sql.query(queryString,[answersData.userName,answersData.questionId],(err,result) => 
-            {
-                if(err || result.length == 0)
-                {
-                    console.log("Invalid question ID or User has NOT YET answered this question. please raise a POST REQUEST to add answer");
-                    res.status(400).send({ message: "Invalid question Id or User has NOT YET answered this question. please raise a POST REQUEST to add answer"});   
-                }
-                else
-                {
-                    // USER has answered this question
-                    // lets update the answer with new content.
-                    const queryString2 = 'update Answers SET content = ? where userName =? and questionId =?;';
-                    sql.query(queryString2,[answersData.content,answersData.userName,answersData.questionId],(err,result) => 
-                    {
-                        if(err)
-                        {
-                            console.log("error: ",err)
-                            res.status(500).send({ message: 'This is an internal server error!'});
-                        }
-                        else
-                        {
-                            console.log("Answer for question "+answersData.questionId+" updated SUCCESSFULLY");
-                            res.status(200).send({message:"Answer for question "+answersData.questionId+" updated SUCCESSFULLY"});
-                        }
-                    });
-                }    
-            });   
+            res.status(400).send({message: "Invalid question Id or User has NOT YET answered this question. please raise a POST REQUEST to add answer"});  
+            return;
         }
-    })
+        await dbConnection.query(updateAnswerQueryString,[answersData.content, answersData.userName, answersData.questionId]);
+        res.status(200).send({message: "Answer for question updated SUCCESSFULLY"});
+    }
+    catch (err)
+    {
+        res.status(500).send({message: "This is an internal server error!"});
+    }
 }
-
-// EXTRA USE CASE
-// UPVOTING AN ANSWER
+   
 /*
+EXTRA USE CASE : UPVOTING AN ANSWER
 
 Function : upvoteAnswer
 
@@ -200,186 +187,188 @@ Response of the API with HTTP STATUS codes:
 
 */
 
-const upvoteAnswer = (req,res) =>
+const upvoteAnswer = async (req,res) =>
 {
     const userData=
     {       
         userName : req.body.userName,
         password : req.body.password
-    }
+    };
     const upvoteData =
     {
         userName : req.body.userName,
         answerId : req.params.id
-    }
-    
-    const authenticateQueryString = 'SELECT * FROM UserDetails where userName = ? and password =?;';
-    sql.query(authenticateQueryString,[userData.userName,userData.password],(err,result) => 
+    };
+    try
     {
-        if(err ||result.length == 0 )
+        var userAuthenticationDetails = await dbConnection.query(authenticateUserQueryString, [userData.userName, userData.password]);
+        if (userAuthenticationDetails.length == 0)
         {
-            console.log("Error: INVALID CREDENTIALS FOR LOGIN. PLEASE TRY AGAIN");
-            res.status(401).send({ message: "Error: INVALID CREDENTIALS FOR LOGIN. PLEASE TRY AGAIN"});
+            res.status(401).send({message: "Error: INVALID CREDENTIALS FOR LOGIN. PLEASE TRY AGAIN"});
+            return;
         }
-        else
+        
+        var userDetails = await dbConnection.query(getUserWhoPostedAnswerQueryString, upvoteData.answerId)
+        if (userDetails.length == 0)
         {
-            // now lets find the user who posted the answer
-            // and make sure current user doesnt match with that
-            const queryString = 'SELECT userName FROM Answers where answerId = ?;';
-            sql.query(queryString,upvoteData.answerId,(err,result)=>
-            {
-                if(err ||result.length == 0 )
-                {
-                    console.log("error: INVALID ANSWER ID ")
-                    res.status(400).send({ message: 'This is an error! INVALID ANSWER ID'});
-                }
-                else
-                {
-                    const userWhoPostedThisAnswer = result[0].userName;
-                    if (userData.userName == userWhoPostedThisAnswer)
-                    {
-                        console.log("ERROR: User trying to UPVOTE HIS OWN ANSWER.");
-                        res.status(403).send({ message: "ERROR: User trying to UPVOTE HIS OWN ANSWER."});
-                    }
-                    else
-                    {
-                        //lets insert the upvote into upvotes table
-                        const queryString2 = 'insert into Votes SET ?;';
-                        sql.query(queryString2,upvoteData,(err,result)=>
-                        {
-                            if(err)
-                            {
-                                console.log("ERROR: User trying to UPVOTE this ANSWER more than ONCE");
-                                res.status(403).send({ message: "ERROR: User trying to UPVOTE this ANSWER more than ONCE"});
-                            }
-                            else
-                            {
-                                //console.log("We have successfully upvoted the answer")
-                                // now lets increment points for the user who posted this answer
-                                const queryString3 = 'insert into UserPoints(userName) values(?);';
-                                sql.query(queryString3,userWhoPostedThisAnswer,(err,result)=>
-                                {
-                                    if(err )
-                                    {
-                                        console.log("Error:",err)
-                                        res.status(500).send({ message: 'This is an internal server error!'});
-                                    }
-                                    else
-                                    {
-                                        console.log("We have successfully upvoted the answer")
-                                        res.status(201).send({message:"WE HAVE SUCCESSFULLY UPVOTED the answer"});
-                                    }
-                                });
-                            }
-                        })
-                    }   
-
-                }
-            
-            })
+            res.status(400).send({message: "INVALID ANSWER ID"});
+            return;
+        }
+        // Lets find the user who posted the answer and make sure current user doesnt match with that  
+        var userWhoPostedThisAnswer = userDetails[0].userName;
+        if (userData.userName == userWhoPostedThisAnswer)
+        {
+            res.status(403).send({message: "ERROR: User trying to UPVOTE HIS OWN ANSWER."});
+            return;
+        }
+        try
+        {
+            await dbConnection.query(insertUpvoteQueryString, upvoteData);
+        } 
+        catch
+        {
+            res.status(403).send({message: "ERROR: User trying to UPVOTE this ANSWER more than ONCE"});
+            return;
         }  
-    })  
-}
-
-const getPrivilegedUsers = (req,res) =>
-{
-    const queryString = 'select userName,Count(*) as Points from UserPoints Group by userName order by Points desc;';
-    sql.query(queryString,(err,result) => 
+         // Now lets increment points for the user who posted this answer         
+        await dbConnection.query(insertUserPointsQueryString, userWhoPostedThisAnswer);
+        res.status(201).send({message: "We have SUCCESSFULLY UPVOTED the answer"});          
+    }
+    catch (err)
     {
-        if(err)
-        {
-            console.log("error: ",err)
-            res.status(404).send({ message: 'This is an error!'});
-        }
-        else
-        {
-            res.status(200).send(JSON.stringify(result));
-        }
-    });   
-}
-const getVoteCountOfAnswers = (req,res) =>
-{
-    const queryString = 'select Votes.answerId,Answers.content,Count(*) as VoteCount from Votes,Answers where Votes.answerId = Answers.answerId Group by Votes.answerId order by VoteCount desc;';
-    sql.query(queryString,(err,result) => 
-    {
-        if(err)
-        {
-            console.log("error: ",err)
-            res.status(404).send({ message: 'This is an error!'});
-        }
-        else
-        {
-            res.status(200).send(JSON.stringify(result));
-        }
-    });   
+        res.status(500).send({message: "This is an internal server error!"});
+    }                                           
 }
 
+/*
+EXTRA USE CASE : DELEETING AN ANSWER
 
-// API TO DELETE ANSWER
+Function : deleteAnswer
 
-const deleteAnswer = (req,res) =>
+API TYPE : DELETE REQUEST
+
+Description : 
+It serves the DELETE request made from the client for deleting an answer.
+For the given answer Id, we first check if it is valid. Then we make sure 
+the user is not deleting somebody else's answer. We delete the given answer
+from both answers table and votes table. So all the upvotes for this 
+answer also gets deleted.
+We also need to authorize the client making the request.
+
+Input parameters :
+From the request we get,
+1) answerId - The answer Id we want to delete
+2) userName - Email Id of the user
+3) password - password of the user
+
+Response of the API with HTTP STATUS codes:
+200: SUCCESSFULLY deleted answer
+400: 1) Invalid answer ID
+     2) User is trying to delete an answer not posted by him
+401: INVALID CREDENTIALS FOR LOGIN. PLEASE TRY AGAIN
+500: "Internal Server Error "
+
+*/
+
+const deleteAnswer = async (req,res) =>
 {
     const userData =
     {       
         userName : req.body.userName,
         password : req.body.password
-    }
-    const answerId = req.params.id
+    };
+    const answerId = req.params.id;
 
-    const authenticateQueryString = 'SELECT * FROM UserDetails where userName = ? and password =?;';
-    sql.query(authenticateQueryString,[userData.userName,userData.password],(err,result) => 
+    try
     {
-        if(err ||result.length == 0 )
+        var userAuthenticationDetails = await dbConnection.query(authenticateUserQueryString, [userData.userName, userData.password]);
+        if (userAuthenticationDetails.length == 0)
         {
-            console.log("Error: INVALID CREDENTIALS FOR LOGIN. PLEASE TRY AGAIN");
-            res.status(401).send({ message: "Error: INVALID CREDENTIALS FOR LOGIN. PLEASE TRY AGAIN"});
+            res.status(401).send({message: "Error: INVALID CREDENTIALS FOR LOGIN. PLEASE TRY AGAIN"});
+            return;
         }
-        else
+        var answerDetails = await dbConnection.query(getAnswerDetailsQueryString, [answerId, userData.userName]);
+        if (answerDetails.length == 0)
         {
-            const queryString = 'SELECT * FROM Answers where answerId = ? and userName =?;';
-    
-            sql.query(queryString,[answerId,userData.userName],(err,result) => 
-            {
-                if(err || result.length ==0)
-                {
-                    console.log("error: INVALID ANSWER ID OR USER has not posted this answer ")
-                    res.status(404).send({ message: 'This is an error! INVALID ANSWER ID OR USER has not posted this answer'});
-            
-                }
-                else
-                {   // first have to remove from votes table. since it has foreign key referencing answers table;
-                    const deleteQueryString = 'delete from Votes where answerId =?;';
-                    sql.query(deleteQueryString,answerId,(err,result)=>
-                    {
-                        if(err)
-                        {
-                            console.log("error: ",err)
-                            res.status(400).send({ message: 'This is an error!'});
-                        }
-                        else
-                        {
-                            const deleteQueryString2 = 'delete from Answers where answerId =?;';
-                            sql.query(deleteQueryString2,answerId,(err,result)=>
-                            {
-                                if(err)
-                                {
-                                    console.log("error: ",err)
-                                    res.status(400).send({ message: 'This is an error!'});
-                                }
-                                else
-                                {
-                                    console.log("SUCCESSFULLY DELETED ANSWER")
-                                    res.status(200).send({ message: "SUCCESFULLY DELETED ANSWER"});
-                                }
-                            })
-                        }
-                    })    
-                }
-            });   
+            res.status(400).send({message: "Error: INVALID ANSWER ID OR USER has not posted this answer"});
+            return;
         }
-    })
-}   
+        await dbConnection.query(deleteVotesQueryString, answerId);
+        await dbConnection.query(deleteAnswerQueryString, answerId);
+        res.status(200).send({message: "SUCCESFULLY DELETED answer"});
+    }
+    catch (err)
+    {
+        res.status(500).send({message: "This is an internal server error!"});
+    } 
+}     
 
+/*
+EXTRA USE CASE : DISPLAY POINTS OF USERS
+
+Function : getPrivilegedUsers
+
+API TYPE : GET REQUEST
+
+Description : 
+It serves the GET request made from the client for 
+displaying the points of the users who's answers have 
+been upvoted by other users. Each upvote gives one point.
+I have not added points> some amount to give privelege.
+It's just a simple having Count(*) > in the query.
+It displays points of the users sorted in DESC.
+
+Input parameters : NONE
+
+Response of the API with HTTP STATUS codes:
+200: "The points of users listed in JSON FORMAT")
+500: "Internal Server Error "
+
+*/
+const getPrivilegedUsers = async (req,res) =>
+{
+    try
+    {
+        var userPoints = await dbConnection.query(getUserPointsQueryString);
+        res.status(200).send(JSON.stringify(userPoints));
+    }
+    catch (err)
+    {
+        res.status(500).send({message: "This is an internal server error!"});
+    } 
+}
+
+/*
+EXTRA USE CASE : DISPLAY UPVOTE COUNT OF ANSWERS
+
+Function : getUpvoteCountOfAnswers
+
+API TYPE : GET REQUEST
+
+Description : 
+It serves the GET request made from the client for 
+displaying the votes of all the answers that have been
+upvoted before. It displays answers sorted in DESC based on vote count.
+
+Input parameters : NONE
+
+Response of the API with HTTP STATUS codes:
+200: "The answer and its vote count listed in JSON FORMAT")
+500: "Internal Server Error "
+
+*/
+const getUpvoteCountOfAnswers = async (req,res) =>
+{
+    try
+    {
+        var upvoteCount = await dbConnection.query(getVoteCountQueryString);
+        res.status(200).send(JSON.stringify(upvoteCount));
+    }
+    catch (err)
+    {
+        res.status(500).send({message: "This is an internal server error!"});
+    } 
+}
 
 // export all these functions.
-module.exports ={postAnswer,updateAnswer,upvoteAnswer,getPrivilegedUsers,getVoteCountOfAnswers,deleteAnswer}
+module.exports ={updateAnswer,postAnswer,upvoteAnswer,deleteAnswer,getUpvoteCountOfAnswers,getPrivilegedUsers}
