+++ naren/MySqlProject/app/controllers/question_controller.js
@@ -1,13 +1,79 @@
-// FOR QUERYING THE DATABASE
-const sql = require("./db.js")
+// Establishing DataBase connection
+const db = require("../database/db.js")
+const dbConnection= db.getDbConnection()
 
 /*
 
 ************ CONTROLLER FOR ALL QUESTION RELATED REQUESTS ********************
 
+*/
 
-// USE CASE 5 OF CASE STUDY - Part 1
-// GET ALL ANSWERS FOR A GIVEN QUESTION
+// QUERY STRINGS FOR THIS CONTROLLER
+const authenticateUserQueryString = "SELECT * FROM UserDetails WHERE userName = ? AND password = ?;";
+const insertQuestionQueryString = "INSERT INTO Questions SET ?;";
+const getQuestionContentQueryString = "SELECT content FROM Questions WHERE questionId = ?";
+const getAnswerContentQueryString = "SELECT content FROM Answers WHERE questionId = ? ORDER BY answerId";
+const getAllQuestionsQueryString ="SELECT * from Questions ORDER BY questionId;";
+const getAllAnswersQueryString = "SELECT * from Answers ORDER BY answerId;";
+
+/*
+USE CASE 3 OF CASE STUDY : POSTING A NEW QUESTION
+
+Function : postQuestion
+
+API TYPE : POST REQUEST
+
+Description : 
+It serves the POST request made from the client for posting a new 
+question. We insert a new question into the database and reespond
+with the question id generated for the question. We also 
+need to authorize the client making the request. 
+
+Input parameters :
+From the request we get,
+1) questionContent - The content of the question to post.
+2) userName - Email Id of the user
+3) password - password of the user
+
+Response of the API with HTTP STATUS codes:
+201: Question posted Successfully with question ID
+401: INVALID CREDENTIALS FOR LOGIN. PLEASE TRY AGAIN
+500: "Internal Server Error "
+
+*/
+
+const postQuestion = async (req,res) =>
+{
+    const questionData =
+    {
+        userName : req.body.user_Details.userName,
+        content : req.body.question
+    };
+    const userData =
+    {
+        userName : req.body.user_Details.userName,
+        password : req.body.user_Details.password
+    };
+
+    try
+    {
+        var userAuthenticationDetails = await dbConnection.query(authenticateUserQueryString, [userData.userName, userData.password]);
+        if (userAuthenticationDetails.length == 0)
+        {
+            res.status(401).send({message: "Error: INVALID CREDENTIALS FOR LOGIN. PLEASE TRY AGAIN"});
+            return;
+        }
+        var insertQuestionQueryResult = await dbConnection.query(insertQuestionQueryString,questionData);
+        res.status(201).send({message: "Question posted Successfully with ID " + insertQuestionQueryResult.insertId});
+    }
+    catch (err)
+    {
+        res.status(500).send({message: "This is an internal server error!"});
+    }    
+}
+
+/*
+USE CASE 5 OF CASE STUDY - Part 1 : GET ALL ANSWERS FOR A GIVEN QUESTION
 
 Function : getAllAnswersForQuestion
 
@@ -38,143 +104,52 @@ Example
 }
 
 */
-const getAllAnswersForQuestion = (req,res) =>
+const getAllAnswersForQuestion = async (req,res) =>
 {
-    
     const userData=
     {       
         userName : req.body.userName,
         password : req.body.password
-    }
-    // display object made for displaying the result
+    };
     const displayObject =
     {
         "Question" : "",
         "Answers" : []
-    }
-    
+    };
     const questionId = req.params.id;
 
-    const authenticateQueryString = 'SELECT * FROM UserDetails where userName = ? and password =?;';
-    sql.query(authenticateQueryString,[userData.userName,userData.password],(err,result) => 
+    try
     {
-        if( err || result.length == 0 )
+        var userAuthenticationDetails = await dbConnection.query(authenticateUserQueryString, [userData.userName, userData.password]);
+        if (userAuthenticationDetails.length == 0)
         {
-            console.log("Error: INVALID CREDENTIALS FOR LOGIN. PLEASE TRY AGAIN");
-            res.status(401).send({ message: "Error: INVALID CREDENTIALS FOR LOGIN. PLEASE TRY AGAIN"});
+            res.status(401).send({message: "Error: INVALID CREDENTIALS FOR LOGIN. PLEASE TRY AGAIN"});
+            return;
         }
-        else
+        var questionContentDetails = await dbConnection.query(getQuestionContentQueryString, questionId);
+        if (questionContentDetails.length == 0)
         {
-            const queryString = 'select content from Questions where questionId = ?'
-    
-            sql.query(queryString,questionId,(err,result) => 
-            {
-                if((err) || result.length == 0)
-                {
-                    console.log("error: INVALID QUESTION ID ")
-                    res.status(400).send({ message: 'This is an error! INVALID QUESTION ID'});
-                }
-                else
-                {
-                    displayObject.Question = result[0].content;
-                    const queryString2 = 'select content from Answers where questionId = ? order by answerId;'
-                    sql.query(queryString2,questionId,(err,result) => 
-                    {
-                        if(err)
-                        {
-                            console.log("error: ",err)
-                            res.status(500).send({ message: 'This is an internal server error!'});
-                        }
-                        else
-                        {
-                            result = JSON.parse(JSON.stringify(result));
-
-                            for (var index =0;index<result.length;index++)
-                            {
-                                displayObject.Answers.push(result[index].content);
-                            }
-                            console.log(JSON.stringify(displayObject));
-                            res.status(200).send(JSON.stringify(displayObject));
-                        }       
-       
-                    });
-                }
-            });
+            res.status(400).send({message: "This is an error! INVALID QUESTION ID"});
+            return;
         }
-    })
-};
-
-// USE CASE 3 OF CASE STUDY
-// POSTING A NEW QUESTION
-/*
-
-Function : postQuestion
-
-API TYPE : POST REQUEST
-
-Description : 
-It serves the POST request made from the client for posting a new 
-question. We insert a new question into the database and reespond
-with the question id generated for the question. We also 
-need to authorize the client making the request. 
-
-Input parameters :
-From the request we get,
-1) questionContent - The content of the question to post.
-2) userName - Email Id of the user
-3) password - password of the user
-
-Response of the API with HTTP STATUS codes:
-201: Question posted Successfully with question ID
-401: INVALID CREDENTIALS FOR LOGIN. PLEASE TRY AGAIN
-500: "Internal Server Error "
-
-*/
-
-const postQuestion = (req,res) =>
-{
-    const questionData =
-    {
-        userName : req.body.user_Details.userName,
-        content : req.body.question
+        displayObject.Question = questionContentDetails[0].content;
+        var answerContentDetails = await dbConnection.query(getAnswerContentQueryString, questionId);
+        answerContentDetails = JSON.parse(JSON.stringify(answerContentDetails));
+        for (var index =0; index<answerContentDetails.length; index++)
+        {
+            displayObject.Answers.push(answerContentDetails[index].content);
+        }
+        res.status(200).send(JSON.stringify(displayObject));
     }
-    const userData =
+    catch (err)
     {
-        userName : req.body.user_Details.userName,
-        password : req.body.user_Details.password
-    }
+        res.status(500).send({message: "This is an internal server error!"});
+    }  
+}          
 
-    const authenticateQueryString = 'SELECT * FROM UserDetails where userName = ? and password =?;';
-    sql.query(authenticateQueryString,[userData.userName,userData.password],(err,result) => 
-    {
-        if(err ||result.length == 0 )
-        {
-            console.log("Error: INVALID CREDENTIALS FOR LOGIN. PLEASE TRY AGAIN");
-            res.status(401).send({ message: "Error: INVALID CREDENTIALS FOR LOGIN. PLEASE TRY AGAIN"});
-        }
-        else
-        {
-            const queryString = 'INSERT INTO Questions SET ?;';
-            sql.query(queryString,questionData,(err,result) => 
-            {
-                if(err)
-                {
-                    console.log("error:  ",err)
-                    res.status(500).send({ message: 'ERROR: this is an internal server error'});
-                }
-                else
-                {
-                    console.log("Question posted Successfully with ID "+result.insertId);
-                    res.status(201).send({message:"Question posted Successfully with ID "+result.insertId});
-                }
-            })
-        }
-    });
-}
 
-// USE CASE 5 OF CASE STUDY - Part 2
-// GETTING ALL QUESTIONS AND ITS CORRESPONDING ANSWERS
 /*
+USE CASE 5 OF CASE STUDY - Part 2 : GETTING ALL QUESTIONS AND ITS CORRESPONDING ANSWERS
 
 Function : getAllQuestionsAndAnswers 
 
@@ -213,57 +188,39 @@ Example
 }
 
 */
-const getAllQuestionsAndAnswers = (req,res) =>
+const getAllQuestionsAndAnswers = async (req,res) =>
 {
-    // display object made for displaying the result
     const displayObject = {};
 
-    const queryString = 'select * from Questions order by questionId;';
-    sql.query(queryString,(err,result) => 
+    try
     {
-        if(err)
+        var questionContentDetails = await dbConnection.query(getAllQuestionsQueryString);
+        questionContentDetails = JSON.parse(JSON.stringify(questionContentDetails));
+        for (var index =0; index<questionContentDetails.length; index++)
         {
-            console.log("error: ",err)
-            res.status(500).send({ message: 'This is an internal server error!'});
+            // Building the question part of the display object
+            var questionBody = {}
+            questionBody.question = questionContentDetails[index].content;
+            questionBody.answers=[];
+            var questionId = String(questionContentDetails[index].questionId);
+            displayObject[questionId]=questionBody;
         }
-        else
+        var answerContentDetails = await dbConnection.query(getAllAnswersQueryString);
+        answerContentDetails = JSON.parse(JSON.stringify(answerContentDetails));
+        for (var index =0; index<answerContentDetails.length; index++)
         {
-            result = JSON.parse(JSON.stringify(result));
-            for (var index =0;index<result.length;index++)
-            {
-                var questionBody = {}
-                questionBody.question = result[index].content;
-                questionBody.answers=[];
-                var questionId = String(result[index].questionId);
-                // building the question part of the object
-                displayObject[questionId]=questionBody;
-            }
-            const queryString2 = 'select * from Answers order by answerId;';
-            sql.query(queryString2,(err,result) => 
-            {
-                if(err)
-                {
-                    console.log("error: ",err)
-                    res.status(500).send({ message: 'This is an internal server error!'});
-                }
-                else
-                {
-                    result = JSON.parse(JSON.stringify(result));
-                    for (var index =0;index<result.length;index++)
-                    {
-                        var questionId = String(result[index].questionId)
-                        var answerContent = result[index].content;
-                        // collecting all answers for each question
-                        displayObject[questionId].answers.push(answerContent)
-                    }
-                    // collected all questions and answers within display object
-                    console.log(displayObject);
-                    res.status(200).send(JSON.stringify(displayObject));
-                }    
-            });
+            // Collecting all answers for each question  
+            var questionId = String(answerContentDetails[index].questionId);
+            var answerContent = answerContentDetails[index].content;
+            displayObject[questionId].answers.push(answerContent);
         }
-    })
-}
-
-// export all these functions.
+        res.status(200).send(JSON.stringify(displayObject));
+    }
+    catch (err)
+    {
+        res.status(500).send({message: "This is an internal server error!"});
+    } 
+} 
+                        
+// Export the functions
 module.exports ={getAllAnswersForQuestion,postQuestion,getAllQuestionsAndAnswers}
