+++ naren/MySqlProject/app/routes/question_routes.js
@@ -2,7 +2,7 @@
 
 module.exports = app => 
 {
-    const questions = require("../controllers/test_question_controller.js");
+    const questions = require("../controllers/Question_Controller.js");
     const router = require("express").Router();
     
     router.post("/getAllAnswersForQuestion/:id",questions.getAllAnswersForQuestion);
