+++ naren/MySqlProject/server.js
@@ -14,9 +14,9 @@ app.get("/test",(req,res) => {
     res.json({message:"welcome to NodeJS App"})
 });
 
-require("./app/routes/user_routes")(app);
-require("./app/routes/question_routes")(app);
-require("./app/routes/answer_routes")(app);
+require("./app/routes/User_Routes")(app);
+require("./app/routes/Question_Routes")(app);
+require("./app/routes/Answer_Routes")(app);
 
 const PORT = 8080;
 app.listen(PORT,() => {
